Converted dynamic queries to use SQLx query builder

+++ crates/repository/src/folder.rs
@@ -1,71 +1,45 @@
-use std::fmt::Write;
-
 use colette_core::folder::Cursor;
-use sea_query::{Expr, Iden, Order, PostgresQueryBuilder, Query};
-use sea_query_binder::SqlxBinder;
-use sqlx::{postgres::PgRow, PgExecutor};
+use sqlx::{postgres::PgRow, PgExecutor, Postgres, QueryBuilder};
 use uuid::Uuid;
-
-#[allow(dead_code)]
-pub enum Folder {
-    Table,
-    Id,
-    Title,
-    ParentId,
-    UserId,
-    CreatedAt,
-    UpdatedAt,
-}
-
-impl Iden for Folder {
-    fn unquoted(&self, s: &mut dyn Write) {
-        write!(
-            s,
-            "{}",
-            match self {
-                Self::Table => "folders",
-                Self::Id => "id",
-                Self::Title => "title",
-                Self::ParentId => "parent_id",
-                Self::UserId => "user_id",
-                Self::CreatedAt => "created_at",
-                Self::UpdatedAt => "updated_at",
-            }
-        )
-        .unwrap();
-    }
-}
 
 pub async fn select<'a>(
     ex: impl PgExecutor<'a>,
     id: Option<Uuid>,
     user_id: Uuid,
     parent_id: Option<Option<Uuid>>,
-    limit: Option<u64>,
+    limit: Option<i64>,
     cursor: Option<Cursor>,
 ) -> sqlx::Result<Vec<PgRow>> {
-    let mut query = Query::select()
-        .columns([Folder::Id, Folder::Title, Folder::ParentId])
-        .from(Folder::Table)
-        .and_where(Expr::col(Folder::UserId).eq(user_id))
-        .and_where_option(id.map(|e| Expr::col(Folder::Id).eq(e)))
-        .and_where_option(parent_id.map(|e| {
-            e.map_or_else(
-                || Expr::col(Folder::ParentId).is_null(),
-                |e| Expr::col(Folder::ParentId).eq(e),
-            )
-        }))
-        .and_where_option(cursor.map(|e| Expr::col(Folder::Title).gt(e.title)))
-        .order_by((Folder::Table, Folder::Title), Order::Asc)
-        .to_owned();
+    let mut qb =
+        QueryBuilder::<Postgres>::new("SELECT id, title, parent_id FROM folders WHERE user_id = ");
+    qb.push_bind(user_id);
 
-    if let Some(limit) = limit {
-        query.limit(limit);
+    if let Some(id) = id {
+        qb.push(" AND id = ");
+        qb.push_bind(id);
+    }
+    if let Some(parent_id) = parent_id {
+        if parent_id.is_some() {
+            qb.push(" AND parent_id = ");
+            qb.push_bind(parent_id);
+        } else {
+            qb.push(" AND parent_id IS NULL");
+        }
     }
 
-    let (sql, values) = query.build_sqlx(PostgresQueryBuilder);
+    if let Some(cursor) = cursor {
+        qb.push(" title > ");
+        qb.push_bind(cursor.title);
+    }
 
-    sqlx::query_with(&sql, values).fetch_all(ex).await
+    qb.push(" ORDER BY title ASC");
+
+    if let Some(limit) = limit {
+        qb.push(" LIMIT ");
+        qb.push_bind(limit);
+    }
+
+    qb.build().fetch_all(ex).await
 }
 
 pub async fn insert<'a>(
@@ -75,9 +49,7 @@ pub async fn insert<'a>(
     user_id: Uuid,
 ) -> sqlx::Result<Uuid> {
     sqlx::query_scalar!(
-        "INSERT INTO folders (title, parent_id, user_id)
-VALUES ($1, $2, $3)
-RETURNING id",
+        "INSERT INTO folders (title, parent_id, user_id) VALUES ($1, $2, $3) RETURNING id",
         title,
         parent_id,
         user_id
@@ -93,32 +65,34 @@ pub async fn update<'a>(
     title: Option<String>,
     parent_id: Option<Option<Uuid>>,
 ) -> sqlx::Result<()> {
-    let mut query = Query::update()
-        .table(Folder::Table)
-        .value(Folder::UpdatedAt, Expr::current_timestamp())
-        .and_where(Expr::col(Folder::Id).eq(id))
-        .and_where(Expr::col(Folder::UserId).eq(user_id))
-        .to_owned();
+    let mut qb = QueryBuilder::<Postgres>::new("UPDATE folders SET ");
+
+    let mut separated = qb.separated(", ");
 
     if let Some(title) = title {
-        query.value(Folder::Title, title);
+        separated.push("title = ");
+        separated.push_bind_unseparated(title);
     }
     if let Some(parent_id) = parent_id {
-        query.value(Folder::ParentId, parent_id);
+        separated.push("parent_id = ");
+        separated.push_bind_unseparated(parent_id);
     }
 
-    let (sql, values) = query.build_sqlx(PostgresQueryBuilder);
+    separated.push("updated_at = now()");
 
-    sqlx::query_with(&sql, values).execute(ex).await?;
+    qb.push(" WHERE id = ");
+    qb.push_bind(id);
+    qb.push(" AND user_id = ");
+    qb.push_bind(user_id);
+
+    qb.build().execute(ex).await?;
 
     Ok(())
 }
 
 pub async fn delete<'a>(ex: impl PgExecutor<'a>, id: Uuid, user_id: Uuid) -> sqlx::Result<()> {
     sqlx::query!(
-        "DELETE FROM folders
-WHERE id = $1
-AND user_id = $2",
+        "DELETE FROM folders WHERE id = $1 AND user_id = $2",
         id,
         user_id
     )
